Cleans up 2 and split add additional hook for binding pip logic into MediaContainer
